chore: bump dependencies
