+++ apps/cron/package.json
@@ -14,14 +14,14 @@
   },
   "license": "AGPL-3.0",
   "dependencies": {
-    "node-cron": "^3.0.3",
-    "@aws-sdk/client-s3": "3.637.0"
+    "@aws-sdk/client-s3": "3.645.0",
+    "node-cron": "^3.0.3"
   },
   "devDependencies": {
     "@tape.xyz/config": "workspace:*",
     "@tape.xyz/constants": "workspace:*",
     "@tape.xyz/server": "workspace:*",
-    "@types/node": "^22.5.2",
+    "@types/node": "^22.5.4",
     "@types/node-cron": "^3.0.11",
     "tsx": "^4.19.0",
     "typescript": "5.5.4"
