+++ apps/web/package.json
@@ -16,23 +16,23 @@
   "dependencies": {
     "@emoji-mart/react": "^1.1.1",
     "@hookform/resolvers": "^3.9.0",
-    "@irys/sdk": "^0.2.10",
+    "@irys/sdk": "^0.2.11",
     "@lens-protocol/metadata": "^1.2.0",
     "@livepeer/react": "3.1.9",
     "@tailwindcss/aspect-ratio": "^0.4.2",
-    "@tanstack/react-query": "^5.53.3",
+    "@tanstack/react-query": "^5.54.1",
     "axios": "^1.7.7",
     "dayjs": "^1.11.13",
     "emoji-mart": "^5.6.0",
     "filereader-stream": "^2.0.0",
-    "framer-motion": "^11.3.30",
+    "framer-motion": "^11.5.3",
     "graphql": "^16.9.0",
     "idb-keyval": "^6.2.1",
     "interweave": "^13.1.0",
     "js-cookie": "^3.0.5",
     "keen-slider": "^6.8.6",
     "make-plural": "^7.4.0",
-    "next": "14.2.7",
+    "next": "14.2.8",
     "next-themes": "^0.3.0",
     "react": "18.3.1",
     "react-cool-inview": "^3.0.1",
@@ -42,7 +42,7 @@
     "react-use-websocket": "^4.8.1",
     "textarea-caret": "^3.1.0",
     "uuid": "^10.0.0",
-    "viem": "^2.21.1",
+    "viem": "^2.21.2",
     "wagmi": "^2.12.7",
     "zod": "^3.23.8",
     "zustand": "^4.5.5"
@@ -57,13 +57,13 @@
     "@tape.xyz/lens": "workspace:*",
     "@tape.xyz/ui": "workspace:*",
     "@types/js-cookie": "^3.0.6",
-    "@types/node": "22.5.2",
+    "@types/node": "22.5.4",
     "@types/react": "18.3.5",
     "@types/react-dom": "18.3.0",
     "@types/textarea-caret": "^3.0.3",
     "@types/uuid": "^10.0.0",
     "autoprefixer": "^10.4.20",
-    "postcss": "^8.4.42",
+    "postcss": "^8.4.45",
     "tailwindcss": "^3.4.10",
     "typescript": "5.5.4"
   }
