+++ apps/webv3/package.json
@@ -14,24 +14,24 @@
     "format": "prettier --write \"**/*.{ts,tsx,md,json,css}\""
   },
   "dependencies": {
-    "@tanstack/react-query": "^5.53.3",
-    "framer-motion": "^11.3.30",
-    "next": "14.2.7",
+    "@tanstack/react-query": "^5.54.1",
+    "framer-motion": "^11.5.3",
+    "next": "14.2.8",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-virtuoso": "^4.10.1"
   },
   "devDependencies": {
-    "@tanstack/react-query-devtools": "^5.54.0",
+    "@tanstack/react-query-devtools": "^5.54.1",
     "@tape.xyz/config": "workspace:*",
     "@tape.xyz/constants": "workspace:*",
     "@tape.xyz/generic": "workspace:*",
     "@tape.xyz/lens": "workspace:*",
     "@tape.xyz/ui": "workspace:*",
-    "@types/node": "^22.5.2",
+    "@types/node": "^22.5.4",
     "@types/react": "18.3.5",
     "@types/react-dom": "18.3.0",
-    "postcss": "^8.4.42",
+    "postcss": "^8.4.45",
     "tailwindcss": "^3.4.10",
     "typescript": "^5"
   }
