+++ packages/constants/package.json
@@ -12,7 +12,7 @@
   },
   "devDependencies": {
     "@tape.xyz/config": "workspace:*",
-    "@types/node": "^22.5.2",
+    "@types/node": "^22.5.4",
     "typescript": "^5.5.4"
   }
 }
