+++ packages/helpers/generic/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@lens-protocol/metadata": "^1.2.0",
     "axios": "^1.7.7",
-    "viem": "^2.21.1"
+    "viem": "^2.21.2"
   },
   "devDependencies": {
     "@tape.xyz/config": "workspace:*",
