+++ packages/ui/package.json
@@ -22,7 +22,7 @@
     "@radix-ui/react-switch": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.0",
     "@tippyjs/react": "^4.2.6",
-    "framer-motion": "^11.3.30",
+    "framer-motion": "^11.5.3",
     "next-themes": "^0.3.0",
     "react": "^18.3.1",
     "tippy.js": "^6.3.7"
